Refactor .forward() pass to enable parallelism over batch inputs
My previous code had used a for loop over input images to generate predictions, which was easier to write/understand, but (I'm assuming) less efficient.

The current code replaces this for loop by the use of batch operations everywhere. This involves a lot more slicing and reshaping operations which make the code harder to read, but I added comments to hopefully keep this reasonably transparent.

I also found and fixed a pretty huge bug in how I had been multiplying matrices. The root of this was an index mixup in a reshaping operation, but the end result was to make the MPS geometry into (more or less) a random graph over the image pixels. The fact that this extremely non-local geometry still permitted decent classification accuracy is really surprising, and might be important to keep in mind when understanding the general behavior of MPS or more general string-bond state classifiers.