Starting development on a new layout which better handles different modes
Instead of treating __init__(), forward(), etc as a collection of if statements,
I'm structuring the definition and evaluation of MPS's in terms of a small language
which implements a fixed number of basic operations. The pieces of this are:

(1) A means of initializing and storing our parameter tensors
(2) A means of "repackaging" our input into a form which is easy to contract
    with the defining tensors, along with a map for each tensor which says
    how this contraction is actually done
(3) A mapping from our input-contracted tensor to a list of "reducibles",
    which are then contracted with each other to according to some fixed
    (efficient) code

I'm starting development on this in a new branch, but wanted to checkpoint first.