Finished basic realizables, added multiplication routine for general contractables
Realizables are now Pytorch Modules, and composite reducibles are now
contractables. The first change is necessary for properly registering
parameters, and the latter is for simplicity, and because a separate
Reducibles class wasn't giving me anything new. The routine realize()
has been replaced by forward(), because that's just how Modules work.

The multiplication routine for arbitrary contractables makes use of
the presence of a bond string for each contractable, and acts to
contract the 'l' bond of one with the 'r' bond of another. The batch
index is matched between the two, but all other indices remain separate.

I need to see what existing multiplication routines I can replace by
this general one, and also want to rename 'contractables' as 'cores',
since the former is kind of a mouthful...