Wrote low-level merging code and laid out high-level merging and unmerging methods
The workflow for merging and unmerging is that each core (Pytorch Module)
in a MergedLinearRegion may have merge() and/or unmerge() methods, which
inplement the actual tensor operations and return a new list of cores.
The merge and unmerge methods for MergedLinearRegion in turn handle just
the overall control of (un)merging, and delegate to the lower-level methods.

Each high-level method will have two passes, the first being a simple
loop through the cores in our list. The second pass will then go
through and handle pairs of small cores which arose from separate cores
in our list.