Wrote flexible SVD routine and finished low-level unmerging code
Every type of core can now be merged (possibly trivially), and the
MergedX cores can also be unmerged. This splits up each core into
smaller pieces, and I just need to write the higher-level routines
appearing in MergedLinearRegion, which deal with inter-core merging
and unmerging. After that, I should finally have dynamic bond dims.

My flexible SVD routine is really promising for general applications,
as it takes in a tensor and an ein_string-style specification and
returns a split of the tensor into two other tensors. The power here
is that I can split an arbitrary tensor along an arbitrary partition
of its indices in a single line. This drastically simplified writing
the unmerging code, and will probably be quite useful in the future.