Wrote tools for hyperparameter search, fixed issue with open boundaries
I had been having issues with training in the presence of either open
boundary conditions or dynamic bond dimensions, so I decided to better
characterize the performance of the MPS as its hyperparameters change.
This study was limited to "nuisance" hyperparameters that I don't want
to think about later, which here is just the learning rate and the
variance of the random normal terms during initialization.

It turns out open boundary conditions have no fundamental issues with
training, but are more sensitive to the size of the random normal terms.
Choosing random_eye initialization with standard deviations ~10**-10
is ideal for such MPS classifiers, so I've changed the default values
accordingly.

The issue with dynamic bond dimensions seems more significant though,
so I'm going to tackle that next.