Fixed bug that had frozen training when using dynamic bond dimensions
It turns out the issue was that after modifying my cores during merging
and unmerging, the new tensors weren't being properly registered with
the optimizer. This caused the optimizer's step() method to be a no-op
on our MPS cores after the first re-merging step, which was the source
of the training freeze.

I adopted the solution of having two separate module lists during
training, with each one setting the parameters for a different merged
configuration. Care is required to only perform in-place updates on
the tensors within these module lists, since assignment-based updates
don't update the list of tensors stored in our optimizer.