Added bond_list attribute and fixed issue with diverging tensor norms
There had been some capability to track bond dimensions at a low level
of the code (close to where the SVD's are occurring), but there
wasn't any means for a user to see the list of dynamically evolving
bond dimensions. Just added a `bond_list` attribute to our MPS which
keeps an updated list of the bond dimensions between all our cores.

There was also an issue with some of the tensors diverging in norm
over time, which arose from certain cores receiving an unbalanced
number of singular value matrices during unmerging. This didn't
appear to harm the classification performance, but I added some
code to explicitly balance out the norms of all the cores in the MPS.