Added custom routing and user-level singular value attribute
During initialization, the user can now give a path list which
specifies the regions of the input we wish to visit. This was
very simple to implement, but opens up a lot of possibilities for
asking questions about the locality of information transmission
within an MPS, as well as experimenting with string bond states.

In a similar spirit to the bond_list attribute from earlier, users
can now view an sv_list attribute which gives the singular value
spectrum for each bond from the last time it was unmerged. This is
initialized to be -1 everywhere, to make it obvious which bonds
haven't gone through merging/unmerging. Previously this information
had been available, but was a huge pain to access, and I'm hoping
this makes it easier to ask interesting questions about truncation
in our adaptive training protocol.