Added functionality to register user-defined feature maps, tweaked initialization
The new register_feature_map attribute takes in a custom embedding
of a single scalar and adds it to our MPS so that it is applied to
all input data. If the feature map we give is a Pytorch Module with
trainable parameters, these parameters will be updated during training
(provided we call register_feature_map before initializating the optimizer)

I also made some changes with how MPS cores are initialized. The main
functional change here is that during adaptive training, our identity
core slices all begin with a bond dimension of `output_dim`, which
could heuristically be argued to be the minimal bond dimension for
"lossless transmission" of label/class information. This reduces an
implicit hyperparameter that had been floating around. Some code
cleanup was involved in all of this too.