Changed format of input to ProbMPS, fixed a previous bug
The forward and loss functions for ProbMPS and ProbUnifMPS had
previously expected input with a spatial dimension first, then a
batch dimension. This is in line with the Pytorch format for RNNs,
but is confusing and introduced some bugs in the code. The current
code only accepts inputs with batch dimensions first.